requirements.txt
python-telegram-bot==21.2
pymongo==4.7.2
pytest-asyncio
pytest~=8.2.1
openpyxl~=3.1.2
setuptools~=68.2.0
python-dotenv==0.20.0
mongoengine==0.28.2